MultilanguageCountryWindowFilter -- fixed dialect shifting logic

+++ business-logic/src/main/java/com/netflix/vms/transformer/modules/countryspecific/VMSAvailabilityWindowModule.java
@@ -149,7 +149,7 @@ public class VMSAvailabilityWindowModule {
                 boolean isAvailableForDownload = entry.getValue().isAvailableForDownload;
 
                 if(rightsContract != null) {
-                    int contractAvailability = locale == null ? -1 : multilanguageCountryWindowFilter.contractAvailabilityForLanguage(locale, rightsContract);
+                    long contractAvailability = locale == null ? -1 : multilanguageCountryWindowFilter.contractAvailabilityForLanguage(locale, rightsContract);
                     
                     ListOfRightsContractPackageHollow packageIdList = rightsContract._getPackages();
